+++ src/components/Page/Page.vue
@@ -1,0 +1,31 @@
+<template>
+  <b-container class="bv-example-row pt-4">
+    <template v-if="allPagesLoaded">
+      <h1>{{ pageContent.title.rendered }}</h1>
+      <div v-html="pageContent.content.rendered"></div>
+    </template>
+    <Loader v-else />
+  </b-container>
+</template>
+
+<script>
+import Loader from '../partials/Loader.vue';
+import { mapGetters } from 'vuex';
+
+export default {
+  computed: {
+    ...mapGetters({
+      page: 'page',
+      allPagesLoaded: 'allPagesLoaded'
+    }),
+
+    pageContent() {
+      return this.page(this.$route.params.pageSlug)
+    }
+  },
+
+  components: {
+    Loader
+  }
+}
+</script>