+++ src/components/partials/Header.vue
@@ -3,13 +3,12 @@
 
         <b-nav-toggle target="nav_collapse"></b-nav-toggle>
 
-        <b-navbar-brand href="#">Vuejs Wordpress Theme Starter</b-navbar-brand>
+        <b-navbar-brand to="/">Vuejs Wordpress Theme Starter</b-navbar-brand>
 
         <b-collapse is-nav id="nav_collapse">
 
             <b-nav is-nav-bar>
             <b-nav-item href="#">Link</b-nav-item>
-            <b-nav-item href="#" disabled>Disabled</b-nav-item>
             </b-nav>
 
             <!-- Right aligned nav items -->
@@ -19,13 +18,6 @@
                 <b-form-input size="sm" class="mr-sm-2" type="text" placeholder="Search"/>
                 <b-button size="sm" class="my-2 my-sm-0" type="submit">Search</b-button>
             </b-nav-form>
-
-            <b-nav-item-dropdown text="Lang" right>
-                <b-dropdown-item href="#">EN</b-dropdown-item>
-                <b-dropdown-item href="#">ES</b-dropdown-item>
-                <b-dropdown-item href="#">RU</b-dropdown-item>
-                <b-dropdown-item href="#">FA</b-dropdown-item>
-            </b-nav-item-dropdown>
 
             <b-nav-item-dropdown right>
                 <!-- Using button-content slot -->