+++ src/components/partials/Loader.vue
@@ -1,0 +1,3 @@
+<template>
+  <div>Loading...</div>
+</template>