+++ src/components/widgets/Pages.vue
@@ -2,7 +2,9 @@
   <div class="widget pages">
     <h3><slot></slot></h3>
     <ul v-if="allPagesLoaded">
-      <li v-for="page in somePages(limit)" :key="page.id">{{ page.title.rendered }}</li>
+      <li v-for="page in somePages(limit)" :key="page.id">
+        <router-link :to="'/'+page.slug">{{ page.title.rendered }}</router-link>
+      </li>
     </ul>
     <div v-else>
       Loading...