+++ src/router/index.js
@@ -2,7 +2,9 @@ import Vue from 'vue'
 import Router from 'vue-router'
 
 // Components
-import Home from '../components/Home'
+import Home from '../components/Home.vue'
+import Post from '../components/Post/Post.vue'
+import Page from '../components/Page/Page.vue'
 
 Vue.use(Router)
 
@@ -11,12 +13,25 @@ const router = new Router({
     {
       path: '/',
       name: 'Home',
-      component: Home,
-      // props: { pageContentID: 383 }
+      component: Home
+    },
+    {
+      // Assuming you're using the default permalink structure for posts
+      path: '/:year/:month/:day/:postSlug',
+      name: 'Post',
+      component: Post
+    },
+    {
+      path: '/:pageSlug',
+      name: 'Page',
+      component: Page
     }
   ],
   mode: 'history',
   base: '',
+
+  // Prevents window from scrolling back to top
+  // when navigating between components/views
   scrollBehavior (to, from, savedPosition) {
     if (savedPosition) {
       return savedPosition